Check, if m_process actually exists in Session::interrupt()

+++ src/backends/julia/juliasession.cpp
@@ -147,7 +147,7 @@ void JuliaSession::interrupt()
     if (expressionQueue().isEmpty())
         return;
 
-    if (m_process->pid())
+    if (m_process && m_process->pid())
     {
         disconnect(m_process, &QProcess::errorOccurred, this, &JuliaSession::reportServerProcessError);
         m_process->kill();
